requirements.txt
Flask
Flask-SQLAlchemy
pandas
numpy
tensorflow
scikit-learn
joblib
PyMySQL
python-dotenv